backend/requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
Flask-SQLAlchemy>=3.1.0
reportlab>=4.0.0
pypdf>=3.17.0
google-genai
python-dotenv>=1.0.0







